D11/Ex1/flink-job/src/main/python/requirements.txt
apache-flink==1.17.1
google-cloud-bigquery==3.31.0
pyflink
protobuf
kafka-python